Window: extract resolution code into ResolutionsPage class
